requirements.txt
matplotlib==3.5.3
numpy==1.23.4
pandas==1.5.1
pandas_profiling==3.4.0
seaborn==0.12.1
streamlit_pandas_profiling==0.1.3
Pillow=9.3.0